requirements.txt
# Core dependencies
aiohttp>=3.9.0
aiofiles>=23.2.0
click>=8.1.0
httpx>=0.25.0
dnspython>=2.4.2
pydantic>=2.5.0
python-dotenv>=1.0.0

# Optional LLM integration
ollama>=0.2.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Data handling
pandas>=2.1.0
